Move Extension functions and Helpers to Domain folder
