+++ core/src/main/java/com/lucas/core/domain/extensions/CurrencyExtensions.kt
@@ -1,4 +1,4 @@
-package com.lucas.core.utils.extensions
+package com.lucas.core.domain.extensions
 
 import com.lucas.core.data.models.CurrencyType
