+++ core/src/main/java/com/lucas/core/domain/extensions/TradingPlatformsExtensions.kt
@@ -1,4 +1,4 @@
-package com.lucas.core.utils.extensions
+package com.lucas.core.domain.extensions
 
 import com.lucas.core.data.models.ExchangePlatformType
